requirements.txt
# ToolHub - Python Dependencies
# FastAPI Backend Requirements

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Image to PDF OCR Tool (Scan2PDF)
pytesseract>=0.3.10
Pillow>=10.1.0
pdf2image>=1.16.3
img2pdf>=0.5.1
PyPDF2>=3.0.1

# Markdown to PDF Tool (DocuMark)
markdown>=3.5.1
weasyprint>=60.1
Pygments>=2.17.2

# Data Validator Tool (DataValidator)
PyYAML>=6.0.1
tomli>=2.0.1; python_version < "3.11"
tomli-w>=1.0.0

# Color Palette Tool (ColorPalette)
# Uses Pillow (already included)

# Utilities
werkzeug>=3.0.1
